+++ shoop/core/models/__init__.py
@@ -19,7 +19,7 @@ from ._categories import Category, CategoryStatus, CategoryVisibility
 from ._configurations import ConfigurationItem
 from ._contacts import (
     AnonymousContact, CompanyContact, Contact, ContactGroup, Gender,
-    get_person_contact, PersonContact
+    get_company_contact, get_person_contact, PersonContact
 )
 from ._counters import Counter, CounterType
 from ._manufacturers import Manufacturer
@@ -80,6 +80,7 @@ __all__ = [
     "CustomerTaxGroup",
     "CustomPaymentProcessor",
     "FixedCostBehaviorComponent",
+    "get_company_contact",
     "get_person_contact",
     "Gender",
     "ImmutableAddress",